+++ pages/listIngredients.php
@@ -19,6 +19,8 @@ $defCatClass = $settings['defCatClass'];
       <div class="dropdown-menu dropdown-menu-right">
         <a class="dropdown-item popup-link" href="/pages/mgmIngredient.php">Add new ingredient</a>
         <a class="dropdown-item" id="csv_export" href="/pages/export.php?format=csv&kind=ingredients">Export to CSV</a>
+        <a class="dropdown-item" id="csv_export" href="/pages/export.php?format=json&kind=ingredients">Export to JSON</a>
+
         <a class="dropdown-item" href="#" data-toggle="modal" data-target="#csv_import">Import from CSV</a>
         <div class="dropdown-divider"></div>
         <a class="dropdown-item" href="#" data-toggle="modal" data-target="#pv_online_import">Import from PV Online</a>